Refactored the MainActivity
Created the NotificationApp-class, put the createNotificationChannel-method there and called it in onCreate(). Created the NotificationStore-class that creates a notification-builder.
