+++ app/src/main/java/com/example/notificationsapp/MainActivity.kt
@@ -1,11 +1,7 @@
 package com.example.notificationsapp
 
 import android.Manifest
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
 import android.content.pm.PackageManager
-import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -19,33 +15,16 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.app.ActivityCompat
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
-import com.example.notificationsapp.common.CHANNEL_DESCRIPTION
-import com.example.notificationsapp.common.CHANNEL_NAME
 import com.example.notificationsapp.common.NOTIFICATION_ID
+import com.example.notificationsapp.notifications.NotificationsStore
 import com.example.notificationsapp.ui.theme.NotificationsAppTheme
 
 class MainActivity : ComponentActivity() {
 
+    private val context = this
+    private fun showNotification(notificationBuilder: NotificationCompat.Builder) {
 
-    private fun createNotificationChannel() {
-        // Create the NotificationChannel, but only on API 26+ because
-        // the NotificationChannel class is not in the Support Library.
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val name = CHANNEL_NAME
-            val importance = NotificationManager.IMPORTANCE_DEFAULT
-            val channel = NotificationChannel(CHANNEL_NAME, name, importance).apply {
-                description = CHANNEL_DESCRIPTION
-            }
-            // Register the channel with the system.
-            val notificationManager: NotificationManager =
-                getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(channel)
-        }
-    }
-
-    private fun showNotification(context: Context) {
-
-        with(NotificationManagerCompat.from(this)) {
+        with(NotificationManagerCompat.from(context)) {
             if (ActivityCompat.checkSelfPermission(
                     context,
                     Manifest.permission.POST_NOTIFICATIONS
@@ -60,22 +39,14 @@ class MainActivity : ComponentActivity() {
                 // for ActivityCompat#requestPermissions for more details.
                 return
             }
-            notify(NOTIFICATION_ID, builder.build())
+            notify(NOTIFICATION_ID, notificationBuilder.build())
         }
     }
 
-    private var builder = NotificationCompat.Builder(this, CHANNEL_NAME)
-        .setSmallIcon(androidx.core.R.drawable.notification_bg_normal)
-        .setContentTitle("My notification")
-        .setContentText("Much longer text that cannot fit one line...")
-        .setStyle(NotificationCompat.BigTextStyle()
-            .bigText("Much longer text that cannot fit one line..."))
-        .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        createNotificationChannel()
-        showNotification(this)
+        val basicNotification = NotificationsStore().getNotification(context)
+        showNotification(basicNotification)
 
         setContent {
             NotificationsAppTheme {
